Backend: layer into api / services / state / infra, split models
Restructure the backend into explicit layers with a one-way dependency rule
(api -> services -> state; both reach the leaves; infra/math at the bottom):

- infra/          snowflake_connection, secret_config, server_timing, auth
- api/            routers (auth/meta/data/sessions/evaluate/trade/platforms),
                  errors, util, schemas (request bodies + router-built responses),
                  platform_helpers
- services/       session_management, pipeline, evaluate, trading
- state/          session store, upload_store
- models.py       shared response DTOs (built by services, declared by api)
- parameters.py, data_retrieval.py  shared leaves

Also: main.py reduced to a composition root (app, logging, middleware,
include_router, static); split data_retrieval's Snowflake plumbing into
infra/snowflake_connection; moved PositionConfig into math/; moved trading out
of math/ (dead-code removed, inner helpers take H not Session); dissolved
helper_functions into its sole consumers; routed all secret reads through
get_secret.

Fixed two __file__-relative path breaks exposed by the moves (secrets.toml,
mean_of_variances.csv). Regenerated the evaluate-signature golden (NBA table
dedup removed duplicate-spelling roster rows; serialization-only, H-scores
unchanged). Full suite: 64 passed.
