+++ backend/api/routers/__init__.py
@@ -1,0 +1,2 @@
+"""FastAPI routers, one module per concern. Each exposes `router` (an APIRouter);
+backend.main includes them into the app."""
