+++ backend/infra/__init__.py
@@ -1,0 +1,5 @@
+"""Domain-free plumbing: Snowflake access, secret resolution, request timing, and OAuth.
+
+These modules would work unchanged if the app stopped being about fantasy basketball —
+that is the criterion for living here rather than at the backend top level.
+"""
