+++ backend/models.py
@@ -1,5 +1,8 @@
 """
-Pydantic request / response models matching api_spec.md.
+Shared response/DTO models — the ones the service tier (evaluate, trading) builds
+and the API tier declares as `response_model`. Both tiers import this, so it lives
+at the backend top level and imports nothing internal. Request bodies and router-only
+responses live in `backend.api.schemas`.
 """
 
 from __future__ import annotations
@@ -8,125 +11,11 @@ from typing import Optional
 from pydantic import BaseModel
 
 
-# ── /data/upload ──────────────────────────────────────────────────────────────
-
-class UploadResponse(BaseModel):
-    data_id: str
-    file_type: str
-    n_players: int
-    expires_at: str
-
-
-# ── /sessions POST ────────────────────────────────────────────────────────────
-
-class LeagueSettings(BaseModel):
-    sport: str
-    n_drafters: int
-    n_picks: int
-    scoring_format: str
-    categories: list[str] = []
-    cash_per_team: Optional[int] = None   # Auction Mode only
-
-
-class ModelParameters(BaseModel):
-    omega: float
-    gamma: float
-    beth: float
-    upsilon: float
-    psi: float
-    chi: float
-    aleph: float
-    n_iterations: int
-    streaming_noise: float
-
-
-class DataSource(BaseModel):
-    type: str
-    season: Optional[str] = None                           # 'historical' type only
-    blend_weights: Optional[dict[str, float]] = None          # 'projections' type only
-    custom_data_ids: Optional[dict[str, Optional[str]]] = None  # 'csv' / 'projections'
-
-
-class PlatformConfigRequest(BaseModel):
-    league_id: str
-    division_id: Optional[str] = None
-
-
-class SessionRequest(BaseModel):
-    league: LeagueSettings
-    platform: str = 'Enter your own data'
-    slot_counts: dict[str, int]
-    parameters: ModelParameters
-    data_source: DataSource
-    injured_players: list[str] = []
-    my_team_id: Optional[str] = None
-    platform_config: Optional[PlatformConfigRequest] = None   # live platforms only
-
-
-class PlayerGScore(BaseModel):
-    name: str
-    total: float
-    values: list[float]   # per-category G-scores, same order as categories
-
-
-class SessionResponse(BaseModel):
-    session_id: str
-    n_players_loaded: int
-    categories: list[str]
-    g_scores: list[PlayerGScore]
-    expires_at: str
-
-
-# ── /sessions/{id} PATCH ──────────────────────────────────────────────────────
-
-class PatchLeague(BaseModel):
-    n_drafters: Optional[int] = None
-    n_picks: Optional[int] = None
-    scoring_format: Optional[str] = None
-    categories: Optional[list[str]] = None
-    cash_per_team: Optional[int] = None
-
-
-class PatchRequest(BaseModel):
-    from_step: int
-    parameters: Optional[ModelParameters] = None
-    league: Optional[PatchLeague] = None
-    data_source: Optional[DataSource] = None
-    slot_counts: Optional[dict[str, int]] = None
-    injured_players: Optional[list[str]] = None
-    platform: Optional[str] = None                            # set when connecting a live platform
-    platform_config: Optional[PlatformConfigRequest] = None   # set when connecting a live platform
-
-
-class PatchResponse(BaseModel):
-    ok: bool
-    steps_rerun: list[int]
-
-
-# ── /sessions/{id}/g-scores GET ───────────────────────────────────────────────
-
-class GScoresResponse(BaseModel):
-    g_scores: list[PlayerGScore]
-
-
 # ── /sessions/{id}/evaluate ───────────────────────────────────────────────────
-
-class EvaluateRequest(BaseModel):
-    player_assignments: dict[str, list[str]]
-    my_team_id: str
-    remaining_cash: Optional[dict[str, float]] = None   # Auction Mode only
-    exclusion_list: list[str] = []
-    # Draft/waiver batching: evaluate only a slice of the candidate pool (ordered by the cached
-    # default/generic H-score ranking) so the top players can paint before the deep bench is scored.
-    # candidate_limit=None evaluates everyone (auction always does; the first eval does too, since it
-    # is what establishes the generic ranking).
-    candidate_offset: int = 0
-    candidate_limit: Optional[int] = None
-
 
 # The candidate expand-view rows below are plain dataclasses rather than Pydantic
 # BaseModels. They are built in bulk (hundreds of candidates × several rows each) by
-# backend.evaluate from already-typed numpy .tolist() output, so per-row Pydantic
+# backend.services.evaluate from already-typed numpy .tolist() output, so per-row Pydantic
 # validation is pure overhead here — dataclass construction is ~4× faster and Pydantic v2
 # still serialises them (nested inside the Candidate/EvaluateResponse models) to identical
 # JSON. The outer Candidate/EvaluateResponse stay Pydantic so the API contract is enforced.
@@ -194,15 +83,6 @@ class EvaluateResponse(BaseModel):
 
 # ── /sessions/{id}/trade/analyze ─────────────────────────────────────────────
 
-class TradeAnalyzeRequest(BaseModel):
-    player_assignments: dict[str, list[str]]
-    my_team: str
-    their_team: str
-    my_trade: list[str]
-    their_trade: list[str]
-    ignore_position_check: bool = False
-
-
 class TeamHScore(BaseModel):
     h_score: float
     rates: list[float]   # per-category win rates
@@ -227,16 +107,6 @@ class ComboParam(BaseModel):
     threshold: float
 
 
-class TradeSuggestRequest(BaseModel):
-    player_assignments: dict[str, list[str]]
-    my_team: str
-    their_team: str
-    combo_params: list[ComboParam]
-    your_differential_threshold: float = 0.0
-    their_differential_threshold: float = -0.20
-    ignore_position_check: bool = False
-
-
 class TradeSuggestion(BaseModel):
     send: list[str]
     receive: list[str]
@@ -246,49 +116,3 @@ class TradeSuggestion(BaseModel):
 
 class TradeSuggestResponse(BaseModel):
     suggestions: list[TradeSuggestion]
-
-
-# ── /platforms/* (live platform integration) ─────────────────────────────────
-
-class DivisionsResponse(BaseModel):
-    divisions: list[dict]            # [{name, id}]; empty when the league has none
-
-
-class LeaguesResponse(BaseModel):
-    leagues: list[dict]              # [{id, name, season}]; empty for manual-id platforms (Fantrax)
-
-
-class ConnectRequest(BaseModel):
-    league_id: str
-    division_id: Optional[str] = None
-
-
-class ConnectResponse(BaseModel):
-    team_names: list[str]
-    n_drafters: int
-    n_picks: int
-    available_modes: list[str]
-
-
-class DraftStateResponse(BaseModel):
-    player_assignments: dict[str, list[str]]
-    injured_players: list[str]
-    status: str
-    remaining_cash: Optional[dict[str, float]] = None   # Auction Mode only
-
-
-# Yahoo OAuth (manual code-paste flow)
-
-class YahooAuthUrlResponse(BaseModel):
-    auth_url: str
-
-
-class YahooTokenRequest(BaseModel):
-    auth_code: str
-
-
-# ESPN auth (s2 + SWID cookies)
-
-class EspnCredentialsRequest(BaseModel):
-    s2: str
-    swid: str
