+++ testing_files/test_and_benchmark_auction.py
@@ -10,8 +10,8 @@ from benchmark_helpers import (
     , _SCORE_TOL
     , _build_session_request
 )
-from backend.session import get_session
-from backend.evaluate import run_evaluate
+from backend.state.session import get_session
+from backend.services.evaluate import run_evaluate
 
 
 def test_evaluate_auction():
