+++ testing_files/test_and_benchmark_draft.py
@@ -24,8 +24,8 @@ from benchmark_helpers import (
     , _ALL_CATEGORIES
     , _build_session_request
 )
-from backend.session import get_session
-from backend.evaluate import run_evaluate
+from backend.state.session import get_session
+from backend.services.evaluate import run_evaluate
 
 # Snake-draft first round: pick i goes to Team (i+1), in this order.
 _FIRST_ROUND_PICKS = [
